test(corpus): attribute daily-driver semantics
